# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HonMoon API",
    "description": "HonMoon API 문서",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/user-activities": {
      "post": {
        "tags": [
          "User Activity"
        ],
        "summary": "활동 기록 생성",
        "description": "사용자가 특정 장소에서 활동했을 때 활동 내역을 기록합니다. (userId당 placeId는 1개만 허용)",
        "operationId": "createActivity",
        "parameters": [
          {
            "name": "placeId",
            "in": "query",
            "description": "장소 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "description",
            "in": "query",
            "description": "활동 설명",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "한강공원에서 산책"
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivityResponse"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "placeId": 10,
                      "description": "한강공원에서 산책"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-activities/{id}": {
      "get": {
        "tags": [
          "User Activity"
        ],
        "summary": "사용자 활동 상세 조회",
        "description": "특정 사용자 활동의 상세 정보를 조회합니다.",
        "operationId": "getUserActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "사용자 활동 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivityResponse"
                },
                "examples": {
                  "success": {
                    "description": "success",
                    "value": {
                      "id": 1,
                      "userId": "123e4567-e89b-12d3-a456-426614174000",
                      "placeId": 10,
                      "description": "한강공원에서 산책",
                      "isCompleted": true,
                      "pointsEarned": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user-activities/place/{placeId}": {
      "get": {
        "tags": [
          "User Activity"
        ],
        "summary": "장소별 활동 내역 조회",
        "description": "특정 장소에서 발생한 모든 활동 내역을 조회합니다.",
        "operationId": "getActivityHistoryByPlace",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "장소 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-activities/me": {
      "get": {
        "tags": [
          "User Activity"
        ],
        "summary": "내 활동 내역 조회",
        "description": "현재 로그인한 사용자의 활동 내역을 조회합니다.",
        "operationId": "getMyActivityHistory",
        "parameters": [
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-activities/me/recent": {
      "get": {
        "tags": [
          "User Activity"
        ],
        "summary": "내 최근 활동 조회",
        "description": "현재 로그인한 사용자의 최근 활동 내역을 조회합니다.",
        "operationId": "getMyRecentActivity",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "조회할 활동 수",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 10
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserPrincipal": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "provider",
          "roles",
          "subject"
        ]
      },
...
      "AnswerCheckResult": {
        "type": "object",
        "properties": {
          "isCorrect": {
            "type": "boolean"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "reasoning": {
            "type": "string"
          },
          "extractedText": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "confidence",
          "isCorrect",
          "reasoning"
        ]
      },
      "UserActivityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "pointsEarned": {
            "type": "integer",
            "format": "int32"
          },
          "textAnswer": {
            "type": "string"
          },
          "selectedChoiceIndex": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedImageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "aiResult": {
            "$ref": "#/components/schemas/AnswerCheckResult"
          },
          "alreadyExists": {
            "type": "boolean"
          }
        },
        "required": [
          "alreadyExists",
          "createdAt",
          "id",
          "isCompleted",
          "modifiedAt",
          "placeId",
          "pointsEarned",
          "userId"
        ]
      },
...
      "ResponseBodyListUserActivityResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserActivityResponse"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
...
    }
  }
}